changes made to program structure

+++ Capgemini Project/CabBookingApplication/CabBookingApplication/src/main/java/com/cg/mts/repository/ITripBookingRepository.java
@@ -1,5 +1,6 @@
 package com.cg.mts.repository;
 
+import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,6 +8,10 @@ import org.springframework.stereotype.Repository;
 import com.cg.mts.entities.TripBooking;
 
 @Repository
-public interface ITripBookingRepository extends JpaRepository<TripBooking, Long> {
-
+public interface ITripBookingRepository extends JpaRepository<TripBooking, Integer>{
+//	public TripBooking insertTripBooking(TripBooking tripBooking);
+//	public TripBooking updateTripBooking(TripBooking tripBooking);
+//	public TripBooking deleteTripBooking(TripBooking tripBooking);
+//	public List<TripBooking> viewAllTripsCustomer(int customerId);
+//	public TripBooking calculateBill(int customerId);
 }
